Use 16-bit signed normalized integers for positions and 32-bit indices
This is the only 16-bit integer type supported for ray tracing geometry
and the 16-bit floating point conversion has some precision issues at
values larger than 256, even though float16 should support all integers
in the range 0..2048 exactly.
Since it is a normalized type (in the range -1..1), we have to scale the
positions according to the number of bits we want to use for the
positions.

+++ res/org/lwjgl/demo/vulkan/closesthit.glsl
@@ -4,7 +4,6 @@
  */
 #version 460
 #extension GL_NV_ray_tracing : require
-#extension GL_EXT_shader_16bit_storage : require
 
 layout(location = 0) rayPayloadInNV Payload {
     vec3 normal;
@@ -14,7 +13,7 @@ layout(location = 0) rayPayloadInNV Payload {
 hitAttributeNV vec2 attribs;
 
 layout(binding = 3, set = 0) buffer Normals { uint n[]; } normals;
-layout(binding = 4, set = 0) buffer Indices { uint16_t i[]; } indices;
+layout(binding = 4, set = 0) buffer Indices { uint i[]; } indices;
 
 void main() {
     ivec3 index = ivec3(
